improved rendering engine; added benchmark class to calculate Q; GPU and CPU engines render volumes differently leading to error later

+++ include/typedefs.h
@@ -102,7 +102,7 @@ struct voxel {
     int x;
     int y;
     int z;
-    int I;
+    double I;
 };
 
 #endif
